requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-jose
passlib[bcrypt]
python-multipart